requirements.txt
numpy
pandas
pytorch
torchvision
mlflow
scikit-learn